% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrndmsg.R
\name{gnrndmsg}
\alias{gnrndmsg}
\title{Generate Random Message to Show of}
\usage{
gnrndmsg(n = 1)
}
\arguments{
\item{n}{How many messages to create}
}
\value{
string vector
}
\description{
Generate Random Message to Show of
}
\examples{

gnrndmsg()
gnrndmsg(1)
gnrndmsg(3)
gnrndmsg(n = 1)

}
