% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_admin.R
\name{kafka_admin}
\alias{kafka_admin}
\title{Constructor for kafka_class_producer}
\usage{
kafka_admin()
}
\description{
Constructor for kafka_class_producer
}
\examples{

producer <- kafka_admin()

}
\seealso{
\link{kafka_admin_class}
}
