% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_admin_class.R
\name{kafka_admin_class}
\alias{kafka_admin_class}
\title{R6 Class for Kafka Admin}
\description{
R6 Class for Kafka Admin

R6 Class for Kafka Admin
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{java_admin}}{reference to the underlying Java object
Initialize}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{kafka_admin_class$new()}}
\item \href{#method-finalize}{\code{kafka_admin_class$finalize()}}
\item \href{#method-start}{\code{kafka_admin_class$start()}}
\item \href{#method-end}{\code{kafka_admin_class$end()}}
\item \href{#method-restart}{\code{kafka_admin_class$restart()}}
\item \href{#method-running}{\code{kafka_admin_class$running()}}
\item \href{#method-props}{\code{kafka_admin_class$props()}}
\item \href{#method-topics_list}{\code{kafka_admin_class$topics_list()}}
\item \href{#method-topics_create}{\code{kafka_admin_class$topics_create()}}
\item \href{#method-topics_delete}{\code{kafka_admin_class$topics_delete()}}
\item \href{#method-clone}{\code{kafka_admin_class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new admin object.
Instead of \code{kafka_admin_class$new()} one can use \code{kafka_admin()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
returns object reference for method chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Code run when object is removed from session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
Spin up admin and connect it to Kafka cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$start()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
returns object reference for method chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end"></a>}}
\if{latex}{\out{\hypertarget{method-end}{}}}
\subsection{Method \code{end()}}{
Disconnect admin from Kafka cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$end()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
returns object reference for method chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restart"></a>}}
\if{latex}{\out{\hypertarget{method-restart}{}}}
\subsection{Method \code{restart()}}{
Disconnect and reconnect admin from Kafka cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$restart()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
returns object reference for method chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-running"></a>}}
\if{latex}{\out{\hypertarget{method-running}{}}}
\subsection{Method \code{running()}}{
Whether or not admin is active (has been started or not)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$running()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE/FALSE
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-props"></a>}}
\if{latex}{\out{\hypertarget{method-props}{}}}
\subsection{Method \code{props()}}{
Retrieving current current set of properties.
If properties are supplied via props parameter thos properties will
be set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$props(..., .properties = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{a series of properties provided as \code{key = "values"}}

\item{\code{.properties}}{a list of properties provided as  \code{.properties = list(key = "values", ...)}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a list of properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topics_list"></a>}}
\if{latex}{\out{\hypertarget{method-topics_list}{}}}
\subsection{Method \code{topics_list()}}{
Query and return list of topics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$topics_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
returns a character vector of topics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topics_create"></a>}}
\if{latex}{\out{\hypertarget{method-topics_create}{}}}
\subsection{Method \code{topics_create()}}{
Create new topics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$topics_create(topic, partition = 1, replication_factor = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topic}}{the topic names to create}

\item{\code{partition}}{the topics number of partitions}

\item{\code{replication_factor}}{the topics replication factor}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a character vector of topics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topics_delete"></a>}}
\if{latex}{\out{\hypertarget{method-topics_delete}{}}}
\subsection{Method \code{topics_delete()}}{
Delete topics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$topics_delete(topic)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topic}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a character vector of topics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_admin_class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
