% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_consumer.R
\name{kafka_consumer}
\alias{kafka_consumer}
\title{Constructor for kafka_class_consumer}
\usage{
kafka_consumer()
}
\description{
Constructor for kafka_class_consumer
}
\examples{

consumer <- kafka_consumer()

}
