% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_get_log_level.R
\name{kafka_get_log_level}
\alias{kafka_get_log_level}
\title{Retrieving Current Java Log Level}
\usage{
kafka_get_log_level()
}
\description{
Retrieving Current Java Log Level
}
\examples{

kafka_get_log_level()


}
