% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_producer.R
\name{kafka_producer}
\alias{kafka_producer}
\title{Constructor for kafka_class_producer}
\usage{
kafka_producer()
}
\description{
Constructor for kafka_class_producer
}
\examples{

producer <- kafka_producer()

}
\seealso{
\link{kafka_producer_class}
}
