% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_records.R
\name{kafka_records}
\alias{kafka_records}
\title{Constructor for kafka_records_class}
\usage{
kafka_records(parent)
}
\arguments{
\item{parent}{enclosing consumer object}
}
\description{
Constructor for kafka_records_class
}
