% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_set_log_level.R
\name{kafka_set_log_level}
\alias{kafka_set_log_level}
\title{kafka_set_log_level}
\usage{
kafka_set_log_level(log_level = NULL)
}
\arguments{
\item{log_level}{[character] indicating the the log level to use:
TRACE, DEBUG, INFO, WARN, ERROR}
}
\description{
kafka_set_log_level
}
\examples{

kafka_set_log_level("info")


}
