% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaesque-package.R
\docType{package}
\name{kafkaesque}
\alias{kafkaesque}
\title{The 'kafkaesque' Package}
\description{
R bindings to the producer, admin and consumer Kafka APIs - see
class descriptions to find out more.
}
\seealso{
\link{kafka_admin_class}, \link{kafka_producer_class}, \link{kafka_consumer_class} as well as the packages extensive README
}
