% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_method_call.R
\name{test_method_call}
\alias{test_method_call}
\title{test_method_call}
\usage{
test_method_call()
}
\description{
Example function that creates an Java class instance and calls a method
of the object  that was created.
}
\examples{

test_method_call()

}
