% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_class_consumer.R
\name{kafka_class_consumer}
\alias{kafka_class_consumer}
\title{R6 Class for Kafka Consumer}
\description{
R6 Class for Kafka Consumer

R6 Class for Kafka Consumer
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cons}}{reference to the underlying Java object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{kafka_class_consumer$new()}}
\item \href{#method-finalize}{\code{kafka_class_consumer$finalize()}}
\item \href{#method-start}{\code{kafka_class_consumer$start()}}
\item \href{#method-end}{\code{kafka_class_consumer$end()}}
\item \href{#method-poll}{\code{kafka_class_consumer$poll()}}
\item \href{#method-subscribe}{\code{kafka_class_consumer$subscribe()}}
\item \href{#method-subscription}{\code{kafka_class_consumer$subscription()}}
\item \href{#method-clone}{\code{kafka_class_consumer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new consumer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
Code run when object is removed from session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\subsection{Method \code{start()}}{
Spin up consumer and connect it to Kafka cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end"></a>}}
\subsection{Method \code{end()}}{
Disconnect consumer from Kafka cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$end()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poll"></a>}}
\subsection{Method \code{poll()}}{
Polling for messages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$poll()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subscribe"></a>}}
\subsection{Method \code{subscribe()}}{
Subscribe to one or more topics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$subscribe(topics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topics}}{character vector defining topics or topic regex to subscribe to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subscription"></a>}}
\subsection{Method \code{subscription()}}{
List current subscription
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$subscription()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
