% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_class_consumer.R
\name{kafka_class_consumer}
\alias{kafka_class_consumer}
\title{R6 Class for Kafka Consumer}
\description{
R6 Class for Kafka Consumer

R6 Class for Kafka Consumer
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{java_consumer}}{reference to the underlying Java object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{kafka_class_consumer$new()}}
\item \href{#method-finalize}{\code{kafka_class_consumer$finalize()}}
\item \href{#method-start}{\code{kafka_class_consumer$start()}}
\item \href{#method-end}{\code{kafka_class_consumer$end()}}
\item \href{#method-poll}{\code{kafka_class_consumer$poll()}}
\item \href{#method-offset}{\code{kafka_class_consumer$offset()}}
\item \href{#method-consume_next}{\code{kafka_class_consumer$consume_next()}}
\item \href{#method-consume_loop_one_by_one}{\code{kafka_class_consumer$consume_loop_one_by_one()}}
\item \href{#method-consume_loop_batch}{\code{kafka_class_consumer$consume_loop_batch()}}
\item \href{#method-props}{\code{kafka_class_consumer$props()}}
\item \href{#method-subscribe}{\code{kafka_class_consumer$subscribe()}}
\item \href{#method-subscription}{\code{kafka_class_consumer$subscription()}}
\item \href{#method-clone}{\code{kafka_class_consumer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new consumer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Code run when object is removed from session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
Spin up consumer and connect it to Kafka cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end"></a>}}
\if{latex}{\out{\hypertarget{method-end}{}}}
\subsection{Method \code{end()}}{
Disconnect consumer from Kafka cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$end()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poll"></a>}}
\if{latex}{\out{\hypertarget{method-poll}{}}}
\subsection{Method \code{poll()}}{
Polling for messages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$poll()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
returns the number of new messages fetched
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-offset"></a>}}
\if{latex}{\out{\hypertarget{method-offset}{}}}
\subsection{Method \code{offset()}}{
If poll() did fetch any messages, they are stored here until the next call to poll()



\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$offset(offset = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consume_next"></a>}}
\if{latex}{\out{\hypertarget{method-consume_next}{}}}
\subsection{Method \code{consume_next()}}{

\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$consume_next()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consume_loop_one_by_one"></a>}}
\if{latex}{\out{\hypertarget{method-consume_loop_one_by_one}{}}}
\subsection{Method \code{consume_loop_one_by_one()}}{
If poll() did fetch any messages, they are stored here until the next
call to poll()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$consume_loop_one_by_one(
  expr = expression(print(msg)),
  check = expression(TRUE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr}}{expression (e.g. via \code{expression(...)}) that will
process a single message}

\item{\code{check}}{expression (e.g. via \code{expression(...)}) that will
evaluate to TRUE or FALSE to either further process messages or
stop processing and ask function to return}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consume_loop_batch"></a>}}
\if{latex}{\out{\hypertarget{method-consume_loop_batch}{}}}
\subsection{Method \code{consume_loop_batch()}}{
If poll() did fetch any messages, they are stored here until the next
call to poll()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$consume_loop_batch(
  expr = expression(print(msgs)),
  check = expression(TRUE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr}}{expression (e.g. via \code{expression(...)}) that will
process a single message}

\item{\code{check}}{expression (e.g. via \code{expression(...)}) that will
evaluate to TRUE or FALSE to either further process messages or
stop processing and ask function to return}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-props"></a>}}
\if{latex}{\out{\hypertarget{method-props}{}}}
\subsection{Method \code{props()}}{
Polling for messages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$props()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subscribe"></a>}}
\if{latex}{\out{\hypertarget{method-subscribe}{}}}
\subsection{Method \code{subscribe()}}{
Subscribe to one or more topics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$subscribe(topics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topics}}{character vector defining topics or topic regex to subscribe to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subscription"></a>}}
\if{latex}{\out{\hypertarget{method-subscription}{}}}
\subsection{Method \code{subscription()}}{
List current subscription
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$subscription()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_class_consumer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
