% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_producer.R
\name{kafka_producer}
\alias{kafka_producer}
\title{kafka_producer}
\usage{
kafka_producer(
  brokerList = "localhost:9092",
  topic = "test",
  type = "undefined",
  keySerializerClass = "org.apache.kafka.common.serialization.StringSerializer",
  valueSerializerClass = "org.apache.kafka.common.serialization.StringSerializer"
)
}
\arguments{
\item{brokerList}{the list of brokers to produce to}

\item{topic}{the topic to use for messages}

\item{type}{the type - think subtopic - of a message}

\item{keySerializerClass}{kafka key serializer class}

\item{valueSerializerClass}{kafka value serializer class}
}
\description{
Initializes and returns object of class = "kafka_producer", "R6".
Messages can be all types of R-objects. These will be serialized to JSON
before sending.
}
\examples{

\dontrun{
 pr <- kafka_producer_new(topic = "test")
 pr$send_message("1", type = "number")
 pr$send_message("2")
 pr$send_message("3")
 pr$send_message(data.frame(a = 1:10, b = "foo"), type = "data")
 pr$send_message(data.frame(a = 1:10, b = "foo"), type = "gubbelgigoo")
}

}
