% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_records.R
\name{kafka_records_class}
\alias{kafka_records_class}
\title{R6 Class for Kafka Records}
\description{
R6 Class for Kafka Records

R6 Class for Kafka Records
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{kafka_records_class$new()}}
\item \href{#method-records_df}{\code{kafka_records_class$records_df()}}
\item \href{#method-next_record}{\code{kafka_records_class$next_record()}}
\item \href{#method-next_record_batch}{\code{kafka_records_class$next_record_batch()}}
\item \href{#method-clone}{\code{kafka_records_class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_records_class$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-records_df"></a>}}
\if{latex}{\out{\hypertarget{method-records_df}{}}}
\subsection{Method \code{records_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_records_class$records_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-next_record"></a>}}
\if{latex}{\out{\hypertarget{method-next_record}{}}}
\subsection{Method \code{next_record()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_records_class$next_record()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-next_record_batch"></a>}}
\if{latex}{\out{\hypertarget{method-next_record_batch}{}}}
\subsection{Method \code{next_record_batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_records_class$next_record_batch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kafka_records_class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
